+++ lib/index.d.ts
@@ -1,2 +1,2 @@
-export { setConfig, wpSiteInfo, wpPostBySlug, wpPublishedPostsSlugsByCategoryId, wpLatestPublishedPostsByCategoryId } from "./functions";
+export { setConfig, wpSiteInfo, wpTestFunction, wpPostBySlug, wpPublishedPostsSlugsByCategoryId, wpLatestPublishedPostsByCategoryId } from "./functions";
 export { formatDate, getTextStringFromHtml } from "./support";