+++ src/index.ts
@@ -1,0 +1,13 @@
+export {
+  setConfig,
+  wpSiteInfo,
+  wpTestFunction,
+  wpPostBySlug,
+  wpPublishedPostsSlugsByCategoryId,
+  wpLatestPublishedPostsByCategoryId
+} from "./functions";
+
+export {
+  formatDate,
+  getTextStringFromHtml
+} from "./support"